requirements.txt
fastapi>=0.68.0
uvicorn[standard]>=0.15.0
pydantic>=1.8.0
fuzzywuzzy>=0.18.0
python-Levenshtein
pandas>=2.0.0
numpy
requests
openai
python-dotenv
openpyxl